PO restore cron task

+++ app/Console/Commands/RestorePONumber.php
@@ -37,8 +37,14 @@ class RestorePONumber extends Command
      */
     public function handle()
     {
+        if (env('AUTOMATIC_UNUSED_PO_RESTORE', false) === true) {
+            return;
+        }
+
         $count = \DB::table('po_track')->where('enabled', '0')->where('created_at', '<=', \DB::raw('DATE_SUB(NOW(), INTERVAL '.env("UNUSED_PO_RESTORE_TIMEOUT", "120").' MINUTE)'))->delete();
-        \Log::info($count." Unused PO's Restored.");
-        echo $count;
+        if($count){
+            \Log::info($count." Unused PO's Restored.");
+            echo $count." Unused PO's Restored.";
+        }
     }
 }
